Adding projects and tests to demonstrate the 200OK problem is the recent implementation (previous commit, db673740269b63f365671ddd3c5aa8406da08c65)
While the implementation is not close to the spec, the StrawberryShakeTests shows that some aspects are still missing.
AuthTests are assumed to be correct, but require a thorough comparison with the spec to ensure their correctness.
That is, the auth tests are questionable - it's possible that a different response should be expected according to the spec.

In the case of the followSpec=true, a .received.txt is added, to show that the received value doesn't align with the expected value.